test(client): Eth keyserver in browser tests.
