+++ packages/client/jest.setup.js
@@ -1,24 +1,10 @@
 import 'reflect-metadata'
 import { GitRevisionPlugin } from 'git-revision-webpack-plugin'
-import express from 'express'
-import { Debug } from './src/utils/log'
-
-const log = Debug('jest global setup')
+import './test/keyserver'
 
 const pkg = require('./package.json')
 
 export default async () => {
-    const app = express()
-    let c = 0
-    app.get('/key', (_req, res) => {
-        c += 1
-        if (c === 1000) { c = 1 }
-        const hexString = c.toString(16)
-        const privkey = '0x' + hexString.padStart(64, '0')
-        log('key endpoint called, returning key ' + privkey)
-        res.send(privkey)
-    })
-    app.listen(45454)
 
     // eslint-disable-next-line
     require('reflect-metadata')
