+++ packages/client/package.json
@@ -94,6 +94,7 @@
         "buffer": "^6.0.3",
         "bytes": "^3.1.1",
         "chromedriver": "^96.0.0",
+        "cors": "^2.8.5",
         "crypto-browserify": "^3.12.0",
         "eslint": "^7.32.0",
         "eslint-config-airbnb": "^18.2.1",
