+++ packages/client/test/browser/realtime.js
@@ -21,15 +21,16 @@ describe('StreamrClient Realtime', () => {
             .click('button[id=connect]')
             .assert.containsText('#result', 'connected')
             .click('button[id=create]')
-            .pause(6000)
+            .pause(12000)
             .assert.containsText('#result', streamName)
             .click('button[id=permissions]')
+            .pause(12000)
             .assert.containsText('#result', 'stream_subscribe')
             .assert.containsText('#result', 'stream_get')
             .click('button[id=subscribe]')
             .assert.containsText('#result', 'subscribed')
             .click('button[id=publish]')
-            .pause(6000)
+            .pause(12000)
             .verify.containsText('#result', '{"msg":0}')
             .verify.containsText('#result', '{"msg":1}')
             .verify.containsText('#result', '{"msg":2}')
